tweak interface
