+++ modules/channel-provider/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@connext/store",
+  "name": "@connext/channel-provider",
   "version": "3.0.1",
   "description": "Channel Provider module for Connext client",
   "main": "dist/index.js",
