+++ modules/test-runner/package.json
@@ -11,6 +11,7 @@
   "license": "ISC",
   "dependencies": {
     "@connext/cf-core": "3.0.1",
+    "@connext/channel-provider": "3.0.1",
     "@connext/client": "3.0.1",
     "@connext/contracts": "1.0.1",
     "@connext/messaging": "3.0.1",
